splitting up transform even further
