+++ renderapi/transform/__init__.py
@@ -1,2 +1,5 @@
 from .transform import *
+from .AffineModels import *
 from .ThinPlateSpline import *
+from .PolynomialModels import *
+from .utils import *
